requirements.txt
adjustText==0.8.3
Keras==2.6.0
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
statsmodels==0.13.0
xgboost==1.5.0
os==1.6.2
Pillow==8.4.0
tkinter==8.6.11
